redesign MessagesBox.jsx to tsx
