+++ src/components/chat/messageInput/MessageInput.jsx
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import { formatISO } from "date-fns";
 
 const MessageInput = ({ chat, addMessage }) => {
   const [input, setInput] = useState("");
